refactor: rename module `integrations` => `nginx`
