+++ core/src/plugins/kubernetes/ephemeral/config.ts
@@ -17,12 +17,12 @@ import { providerConfigBaseSchema } from "../../../config/provider.js"
 import { ConfigurationError } from "../../../exceptions.js"
 import type { ConfigureProviderParams } from "../../../plugin/handlers/Provider/configureProvider.js"
 import { dedent } from "../../../util/string.js"
-import type { KubernetesConfig, KubernetesPluginContext } from "../config.js"
+import type { KubernetesConfig } from "../config.js"
 import { defaultResources } from "../config.js"
 import { namespaceSchema } from "../config.js"
 import { EPHEMERAL_KUBERNETES_PROVIDER_NAME } from "./ephemeral.js"
 import { DEFAULT_GARDEN_CLOUD_DOMAIN } from "../../../constants.js"
-import { helmNginxInstall, NginxHelmValuesGetter } from "../integrations/nginx.js"
+import { NginxHelmValuesGetter } from "../nginx/nginx.js"
 import type { SystemVars } from "../init.js"
 import { getSystemNamespace } from "../namespace.js"
