+++ core/src/plugins/kubernetes/nginx/nginx.ts
